+++ tests/api_v2/fixtures/002_comparison.h
@@ -2,7 +2,7 @@
     @file
     @author  Alexander Sherikov
 
-    @copyright 2017-2018 Alexander Sherikov, Licensed under the Apache License, Version 2.0.
+    @copyright 2017-2026 Alexander Sherikov, Licensed under the Apache License, Version 2.0.
     (see @ref LICENSE or http://www.apache.org/licenses/LICENSE-2.0)
 
     @brief
@@ -47,7 +47,7 @@ namespace ariles_tests
             ariles2::Compare::Parameters param;
             param.double_tolerance_ = g_tolerance;
             param.compare_number_of_entries_ = true;
-            BOOST_CHECK(ariles2::apply(visitor, configurable_out, configurable_in, param));
+            compareAndCheckWithBacktrace(visitor, configurable_out, configurable_in, param, "ComparisonSimpleFixture failed");
         }
     };
 
@@ -94,8 +94,8 @@ namespace ariles_tests
             ariles2::Compare::Parameters param;
             param.double_tolerance_ = g_tolerance;
             param.compare_number_of_entries_ = true;
-            BOOST_CHECK(ariles2::apply(visitor, configurable_out1, configurable_in1, param));
-            BOOST_CHECK(ariles2::apply(visitor, configurable_out2, configurable_in2, param));
+            compareAndCheckWithBacktrace(visitor, configurable_out1, configurable_in1, param, "ComparisonMultiFixture 1 failed");
+            compareAndCheckWithBacktrace(visitor, configurable_out2, configurable_in2, param, "ComparisonMultiFixture 2 failed");
         }
     };
 }  // namespace ariles_tests